requirements.txt
streamlit
pandas
numpy
matplotlib
reportlab
pyyaml
